requirements.txt
azure-ai-inference
gradio
python-dotenv